feat: add register processing service
